refactor: restructure folders for timeline feature
