+++ src/features/timeline/components/Header.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState } from 'react';
-import Tab from '../../componenets/Tab';
+import Tab from '../../../componenets/ui/home/Tab';
 
 export default function Header() {
   const [selectedTab, setSelectedTab] = useState<null | number>(1);
