+++ src/features/timeline/components/TweetOptionsBar.tsx
@@ -1,4 +1,4 @@
-import Icon from '../../componenets/Icon';
+import Icon from '../../../componenets/ui/home/Icon';
 
 import Schedule from './Schedule';
 export default function TweetOptionsBar({
