+++ src/features/timeline/components/TweetReplySettings.tsx
@@ -1,4 +1,9 @@
-export const options = [
+'use client';
+import { useState } from 'react';
+import Icon from '../../../componenets/ui/home/Icon';
+import useAddTweetStore from '@/features/timeline/store/useAddTweetStore';
+
+const options = [
   {
     id: 1,
     name: 'Everyone',
@@ -24,3 +29,31 @@ export const options = [
     viewBox: 24,
   },
 ];
+export default function TweetReplySettings() {
+  const [replyOption, setReplyOption] = useState(0);
+  const isOpenReplySettings = useAddTweetStore(
+    (state) => state.isOpenReplySettings
+  );
+  if (!isOpenReplySettings) return null;
+  return (
+    <div className="w-full max-h-9 flex flex-1 items-stretch   border-b-1 border-border ">
+      <button
+        onClick={() => {}} // open model
+        className=" h-6 flex items-center justify-center cursor-pointer hover:bg-icon-hover hover:rounded-full text-primary text-sm font-bold pr-3 "
+      >
+        <Icon
+          viewBox={options[replyOption].viewBox}
+          height="h-7"
+          width="w-7"
+          disabled={true}
+          size="w-4"
+          color="text-primary"
+          path={options[replyOption].path}
+        />
+        <span className="text-center font-bold">
+          {options[replyOption].name} can reply
+        </span>
+      </button>
+    </div>
+  );
+}
